+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project does not adhere to [Semantic Versioning](http://semver.org/spec/v2.
 
 ### Fixed
 - Remove redundant default value from `--colour` help text (see [GH-10][]).
+- Replace light black with green; the former doesn't work with Solarized Dark.
 
 [GH-10]: https://github.com/lunaryorn/mdcat/pull/10